Clear up home page learning example
